test(action): SDK Publish Workflow (#188)
This pull request includes several changes to the deployment process for
the JavaScript SDK in the
`.github/actions/core-cicd/deployment/deploy-javascript-sdk/action.yml`
file. The primary focus is on updating the versioning and tagging system
from "alpha" to "beta".

Changes to versioning and tagging:

* Updated the example projects to point to the 'next' tag instead of
'latest' to prevent version inconsistency.
* Changed the default `npm-package-tag` from 'alpha' to 'beta'.
* Modified the script to fetch the current version using the 'beta' tag
instead of 'alpha', and set the initial version to "0.0.1-beta.0" if no
current version is found.
* Adjusted the version calculation logic to increment the beta version
number instead of the alpha number.
* Changed the npm dist-tag addition command to use the 'next' tag
instead of 'latest'.